Replace old yum_src plugin with new zypper plugin
